Refactor
   * Add lombok
   * Add initiial test
