--- src/main/java/net/truedigital/sample/paymentservice/Bundle.java
@@ -1,16 +1,0 @@
-package net.truedigital.sample.paymentservice;
-
-import java.util.List;
-
-public class Bundle {
-    private int discountRate;
-    private List<String> productCodes;
-
-    public int getDiscountRate() { return discountRate; }
-
-    public void setDiscountRate(int discountRate) { this.discountRate = discountRate; }
-
-    public List<String> getProductCodes() { return productCodes; }
-
-    public void setProductCodes(List<String> productCodes) { this.productCodes = productCodes; }
-}
