--- src/main/java/net/truedigital/sample/paymentservice/Product.java
@@ -1,31 +1,0 @@
-package net.truedigital.sample.paymentservice;
-
-import java.math.BigDecimal;
-
-public class Product {
-    private BigDecimal price;
-    private String productCode;
-    private String productDesc;
-
-    public BigDecimal getPrice() { return price; }
-
-    public void setPrice(BigDecimal price) {
-        this.price = price;
-    }
-
-    public String getProductCode() {
-        return productCode;
-    }
-
-    public void setProductCode(String productCode) {
-        this.productCode = productCode;
-    }
-
-    public String getProductDesc() {
-        return productDesc;
-    }
-
-    public void setProductDesc(String productDesc) {
-        this.productDesc = productDesc;
-    }
-}
