+++ src/main/java/net/truedigital/sample/paymentservice/controller/PaymentController.java
@@ -1,5 +1,8 @@
-package net.truedigital.sample.paymentservice;
+package net.truedigital.sample.paymentservice.controller;
 
+import net.truedigital.sample.paymentservice.model.Charge;
+import net.truedigital.sample.paymentservice.model.Payment;
+import net.truedigital.sample.paymentservice.model.Source;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Collections;
@@ -10,22 +13,22 @@ public class PaymentController {
 
     @GetMapping("/accounts/{id}/payments")
     public List<Payment> getPayment(@PathVariable String id) {
-        return Collections.singletonList(new Payment(id, "defaultPaymentName"));
+        return Collections.singletonList(Payment.builder().id(id).name("defaultPaymentName").build());
     }
 
     @GetMapping("/accounts/{id}/sources/{sid}")
     public Source getSource(@PathVariable String sid) {
-        return new Source(sid, "defaultSourceName");
+        return Source.builder().id(sid).build();
     }
 
     @PatchMapping("/accounts/{id}/sources/{sid}")
     public Source patchSource(@PathVariable String sid) {
-        return new Source(sid);
+        return Source.builder().id(sid).build();
     }
 
     @DeleteMapping("/accounts/{id}/sources/{sid}")
     public Source deleteSource(@PathVariable String sid) {
-        return new Source(sid);
+        return Source.builder().id(sid).build();
     }
 
     @PostMapping("/accounts/{id}/sources/{sid}/verify")
@@ -46,7 +49,7 @@ public class PaymentController {
     @GetMapping("/accounts/{id}/charges/{cid}")
     public Charge getCharge(@PathVariable String cid) {
         //TODO: to implement
-        return new Charge(cid);
+        return Charge.builder().id(cid).build();
     }
