+++ src/main/java/net/truedigital/sample/paymentservice/model/Bundle.java
@@ -1,0 +1,17 @@
+package net.truedigital.sample.paymentservice.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class Bundle {
+    private int discountRate;
+    private List<String> productCodes;
+}
