+++ src/main/java/net/truedigital/sample/paymentservice/model/Product.java
@@ -1,0 +1,18 @@
+package net.truedigital.sample.paymentservice.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class Product {
+    private BigDecimal price;
+    private String productCode;
+    private String productDesc;
+}
