--- src/test/java/net/truedigital/sample/paymentservice/PaymentServiceApplicationTests.java
@@ -1,16 +1,0 @@
-package net.truedigital.sample.paymentservice;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class PaymentServiceApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
